+++ generators/kiss99.c
@@ -88,7 +88,7 @@ static int run_self_test(const CallerAPI *intf)
     obj.z   = 12345; obj.w     = 65435;
     obj.jsr = 34221; obj.jcong = 12345; 
     for (long i = 1; i < 1000001 + 256; i++) {
-        val = get_bits_raw(&obj);
+        val = (uint32_t) get_bits_raw(&obj);
     }
     intf->printf("Reference value: %u\n", refval);
     intf->printf("Obtained value:  %u\n", val);
